% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\docType{methods}
\name{checkEnsemblArchive}
\alias{checkEnsemblArchive}
\alias{checkEnsemblArchive,character-method}
\title{Query Ensembl Archive using Ensembl IDs to confirm whether ID is archived}
\usage{
checkEnsemblArchive(query)

\S4method{checkEnsemblArchive}{character}(query)
}
\arguments{
\item{query}{a string containing the ID to be queried.}
}
\value{
"archived" if successful (ID present in Archive), else errors out.
}
\description{
\code{checkEnsemblArchive} contacts the Ensembl Rest API, querying the presence
of an Ensembl ID in the archive database.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: as above.
}}

\examples{
query <- "ENSG00000121031"
ret <- checkEnsemblArchive(query)
}
