% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genematrix}
\alias{genematrix}
\title{Input Matrix of Drug/Gene Data}
\format{A 2 column matrix where the first columns contains drug identifiers and the second column Gene names.}
\usage{
genematrix
}
\description{
A matrix made up of a column of drug identifiers and a column of associated gene names, as used for input into \code{getUniqueGeneList()} and
}
\keyword{datasets}

