% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\docType{methods}
\name{getEnsembleList}
\alias{getEnsembleList}
\alias{getEnsembleList,character-method}
\title{Query Ensembl using Ensembl IDs to find Gene Name}
\usage{
getEnsembleList(query)

\S4method{getEnsembleList}{character}(query)
}
\arguments{
\item{query}{a string containing the ID to be queried.}
}
\value{
A list detailing the entry in the Ensembl database... or "archived" if present in the archive.
}
\description{
\code{getEnsembleList} contacts the Ensembl Rest API, querying the presence
of an Ensembl ID in the active database.
Should the ID not be found in the active database, \code{checkEnsemblArchive()}
is called to check whether it is archived.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: as above.
}}
\examples{
query <- "ENSG00000157764"
ret <- getEnsembleList(query)
ret$display_name
}

