% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnno.R
\docType{methods}
\name{getPublicationList}
\alias{getPublicationList}
\alias{getPublicationList,character-method}
\title{Gather NIH Pubmed Data}
\usage{
getPublicationList(query)

\S4method{getPublicationList}{character}(query)
}
\arguments{
\item{query}{A query string to be submitted to Pubmed}
}
\value{
A vector of \code{pubmed} objects each detailing a publication relating to the entered query combination requested.
}
\description{
\code{getPublicationList} collates Pubmed data returned from a html request for input string.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: as above.
}}
\examples{
\dontrun{
query <- "Thompson IR HIV"
ReturnedPublications <- getPublicationList(query)
}
}

