% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputlist}
\alias{inputlist}
\title{Input Vector of Drug/Gene Data}
\format{A vector of 10 drug identification numbers, intersperced with 90 Gene names.}
\usage{
inputlist
}
\description{
A mixed list of drug identifiers and gene names, as output by \code{parseInputFile()}
and used for input into \code{getUniqueGeneList()} and
}
\keyword{datasets}

