% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mygeneanno}
\alias{mygeneanno}
\title{\code{geneanno} object}
\format{An object of class \code{geneanno} of length 1.}
\usage{
mygeneanno
}
\value{
A geneanno object.
}
\description{
A \code{geneanno} object produced by \code{getUniqueGeneList()} and provided for testing downstream functionality.
}
\keyword{datasets}
