% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mygenelist}
\alias{mygenelist}
\title{Input Vector of Drug/Gene Data}
\format{An object of class \code{character} of length 60.}
\usage{
mygenelist
}
\value{
A vector of 10 drug identification numbers, intersperced with 90 Gene names.
}
\description{
A mixed list of drug identifiers and gene names, based on the genematrix data, as output by \code{parseInputFile()}
and used for input into \code{getUniqueGeneList()}, \code{getGroupGeneList()} and \code{searchPublications()}
}
\keyword{datasets}

