% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mygenematrix}
\alias{mygenematrix}
\title{Input Matrix of Drug/Gene Data}
\format{An object of class \code{matrix} with 50 rows and 2 columns.}
\usage{
mygenematrix
}
\value{
A 2 column matrix where the first columns contains drug identifiers and the second column Gene names.
}
\description{
A matrix made up of a column of drug identifiers and a column of associated gene names, as used for input into \code{getUniqueGeneList()} and
}
\keyword{datasets}
