% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mygenesummaries}
\alias{mygenesummaries}
\title{Vector Input Vector of Drug/Gene Data}
\format{An object of class \code{list} of length 35.}
\usage{
mygenesummaries
}
\value{
A vector of 10 drug identification numbers, intersperced with 90 Gene names.
}
\description{
A mixed list of drug identifiers and gene names, based on the genematrix data, as output by \code{parseInputFile()}
and used for input into \code{getUniqueGeneList()}, \code{getGroupGeneList()} and \code{searchPublications()}
}
\keyword{datasets}
