% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mypublications}
\alias{mypublications}
\title{matrix of \code{pubmed} objects}
\format{An object of class \code{matrix} with 44 rows and 2 columns.}
\usage{
mypublications
}
\value{
A matrix of identifiers from mygenematrix/mygenelist and vectors of \code{pubmed} objects.
}
\description{
A matrix containing the mygenematrix identifiers and  up to 10 related journal articles returned from \code{searchPublications()}
}
\keyword{datasets}

