% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnnoClasses.R
\docType{class}
\name{pubmed-class}
\alias{pubmed}
\alias{pubmed-class}
\title{An S4 class to represent a query to the Pubmed database, inherits from \code{geneanno} class.}
\description{
An S4 class to represent a query to the Pubmed database, inherits from \code{geneanno} class.
}
\section{Slots}{

\describe{
\item{\code{Id}}{The pubmed ID for the publication}

\item{\code{Authors}}{List of article authors}

\item{\code{Date}}{The publication date of the article}

\item{\code{Title}}{Publication Title}

\item{\code{Journal}}{The publication journal of the article}

\item{\code{Volume}}{The journal volume of the article}

\item{\code{Issue}}{The journal volume issue number of the article}

\item{\code{Pages}}{The journal pages of the article}

\item{\code{DOI}}{The unique article Digital Object Identifier (DOI, see www.doi.org)}
}}

