% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnno.R
\docType{class}
\name{query-class}
\alias{query}
\alias{query-class}
\title{An S4 class to represent an initial query to the NCBI e-utilities, inherits from \code{geneanno} class.}
\description{
An S4 class to represent an initial query to the NCBI e-utilities, inherits from \code{geneanno} class.
}
\section{Slots}{

\describe{
\item{\code{gene}}{A character vector of the gene name.}

\item{\code{querykey}}{a character string identifier from the NCBI Eutils.}

\item{\code{webenv}}{A character string uniquely identifying the query from the NCBI Eutils.}
}}

