% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnno.R
\docType{methods}
\name{produceOutputFiles}
\alias{produceOutputFiles}
\alias{produceOutputFiles,geneanno-method}
\title{Save Gene Information to group Specific Output Files}
\usage{
produceOutputFiles(x, dgl, gs)

\S4method{produceOutputFiles}{geneanno}(x, dgl, gs)
}
\arguments{
\item{x}{object of class geneanno.}

\item{dgl}{a Vector of group specific vectors relating the list of genes to the groups, as in the input file}

\item{gs}{Vector of Gene Summary information downloaded from NIH and Uniprot databases, typically output by getGeneSummary}
}
\description{
\code{produceOutputFiles} saves the downloaded gene summary information into group specific files.
}
\section{Methods (by class)}{
\itemize{
\item \code{geneanno}: Save object data to text files.
}}

